Removes the PermutationWorkerFactory.lazyFactories cache.
The problem with caching these as static state is that the Factories themselves do not appear to be stateless.  In particular, ExternalPermutationWorkerFactory has instance state related to accepting a particular number of connections.  This is neither threadsafe and also makes it impossible to reuse.  We can revisit this change if we decide that PWFs should be both threadsafe and reusable (or stateless).

Patch by: scottb, spoon
Review by: bobv (TBR), spoon

git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@5080 8db76d5a-ed1c-0410-87a9-c151d255dfc7