Moving AttributeParsers to XMLElement, where they belong.
This is a pretty major refactoring, and it's not done yet. Management of the
AttributeParser and the BundleAttributeParsers is now in dedicated classes.
XMLElement has access to them, and uses them for Boolean and Double values. All
other parsing is still going through old calls through UiBinderWriter to the new
managers, for now.

Of particular note: there are now many fewer direct dependencies on
UiBinderWriter, and much more DI style injection of logging and other services.
Also, dummy loggers are introduced to eliminate console noise from unit tests.

Next steps are making ComputedAttributeParser use the new game (and in the
process tidy its fix for currently broken {{ escaping), and making the new
LayoutPanelParsers get their enum values in the new style. Then on to strings.

Review by: hermes@google.com

git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@6589 8db76d5a-ed1c-0410-87a9-c151d255dfc7