Improves @UiFactory's handling of generics from catastrophic to inadequate.
We were generating bad code when @UiFactory was put
on a method with a parameterized return type. The
solution here is simplistic, treating everything
as the raw type and not validating that the
param type matches what is expected in the template,
but I think it's an improvement.

The actual fix is small, but the change to the
test classes is pretty big: OwnerClassTest no
longer uses JClassTypeAdapter, which I'm hoping eventually
to delete in favor of mock compile state. This was necessary
because JClassTypeAdapter doesn't deal with generics.

Review by bobv
http://gwt-code-reviews.appspot.com/106801

git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@7032 8db76d5a-ed1c-0410-87a9-c151d255dfc7