Fixes a couple of issues with the RemoteUI and GPE DevMode view interaction.  The problem is GPE shows the startup URLs in the DevMode view _before_ the modules were fully loaded.  The root issue is GPE does not receive the moduleLoadComplete message.  When we went to push this message to GPE, we discovered the protobuf is too restricted:  the RequestType cannot have new values and maintain backwards compatibility.  This is because requestType is a "required" field, so older GPE versions will not understand the new moduleLoadComplete message and thus have an undefined requestType, at which point an error is thrown because requestType is required.
For the protobuf issue, this patch changes the "required" to "optional".
We will still have compatibility issues with GWT or GPE versions
earlier than this patch, but this allows for future versions to have new
values added to the enums.  It is up to the application layer to check
the validity of the message.

For the GPE startup URL issue, it will show a "Loading" state as soon
as DevMode is launched.  Then, GWT will send the startup URLs in
moduleLoadComplete.  Upon receiving these, GPE will display them.  The
user will not see startup URLs before they are ready.

Review at http://gwt-code-reviews.appspot.com/323801

Review by: rdayal@google.com

git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@7931 8db76d5a-ed1c-0410-87a9-c151d255dfc7